fix issues and new styles.

+++ routes/api_v5/clients.php
@@ -3,15 +3,15 @@
 use Illuminate\Support\Facades\Route;
 use App\V5\Modules\Client\Controllers\ClientController;
 
-Route::middleware('auth:sanctum')
+Route::middleware(['auth:sanctum', 'school.context.middleware'])
     ->prefix('clients')
     ->name('v5.clients.')
     ->group(function () {
-        Route::get('/', [ClientController::class, 'index'])->name('index');
-        Route::post('/', [ClientController::class, 'store'])->name('store');
-        Route::get('/{client}', [ClientController::class, 'show'])->name('show');
-        Route::patch('/{client}', [ClientController::class, 'update'])->name('update');
-        Route::delete('/{client}', [ClientController::class, 'destroy'])->name('destroy');
+        Route::get('/', [ClientController::class, 'index'])->middleware('role.permission.middleware:client.read')->name('index');
+        Route::post('/', [ClientController::class, 'store'])->middleware('role.permission.middleware:client.create')->name('store');
+        Route::get('/{client}', [ClientController::class, 'show'])->middleware('role.permission.middleware:client.read')->name('show');
+        Route::patch('/{client}', [ClientController::class, 'update'])->middleware('role.permission.middleware:client.update')->name('update');
+        Route::delete('/{client}', [ClientController::class, 'destroy'])->middleware('role.permission.middleware:client.delete')->name('destroy');
 
         Route::post('/{client}/utilizadores', [ClientController::class, 'storeUtilizador'])->name('utilizadores.store');
         Route::patch('/{client}/utilizadores/{utilizador}', [ClientController::class, 'updateUtilizador'])->name('utilizadores.update');
